Update .gitlab-ci.yml and README.md to specify production Jekyll configuration
If a user is using GitLab Pages it is most likely the user is using GitLab pages to publish a website not just stage a website that is published elsewhere.  If a user is publishing a website with Jekyll, they (and Jekyll themes/plugins) will expect that JEKYLL_ENV=production.  This commit provides that specification in `.gitlab-ci.yml`.

An examination of the existing repository showed this should not result in any change to the current rendering of the website.  However, the specification may be important for developers as explained in preceding commit.